Tests: moved helper method to the individual testsuites to prevent clutter
